requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
facebook-sdk==3.1.0
gunicorn==20.0.4
idna==2.8
numpy==1.18.0
Pillow==6.1.0
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7